+++ app/modules/editor/scenes/EditorEnduranceScene.tsx
@@ -1,6 +1,6 @@
 import React, { FC } from 'react';
-import SceneWrapper from '../../components/SceneWrapper/SceneWrapper';
-import EnduranceForm from './components/EnduranceForm/EnduranceForm';
+import SceneWrapper from '../../../components/SceneWrapper/SceneWrapper';
+import EnduranceForm from '../components/EnduranceForm/EnduranceForm';
 
 const EditorEnduranceScene: FC = () => (
   <SceneWrapper>
